+++ js/scenes/GUIScene.js
@@ -3,31 +3,34 @@ class GUIScene extends Phaser.Scene {
     constructor()
     {
         super({ key: 'GUIScene', active: false})
-        
     }
 
     create()
     {
-        let game = this.scene.get('GameScene');
+        // Grab connect to game scene.
+        this.game = this.scene.get('GameScene');
 
         // Contianer for GUI elements.
         this.elements = [];
 
-        this.guiFont = game.gameConfig.gui.font;
+        this.guiFont = this.game.gameConfig.gui.font;
 
         // Create player hp/mp bars.
-        let bar = game.player.combat.createBar(1, 1, game.gameConfig.gui.bars.width, game.gameConfig.gui.bars.height, 
-                                             game.gameConfig.gui.bars.color.hp, false, 'hp', this);
+        let bar = this.game.player.combat.createBar(1, 1, this.game.gameConfig.gui.bars.width, this.game.gameConfig.gui.bars.height, 
+                                             this.game.gameConfig.gui.bars.color.hp, false, 'hp', this);
 
-        let bar2 = game.player.combat.createBar(game.cameras.main.width-game.gameConfig.gui.bars.width-1, 1, game.gameConfig.gui.bars.width, game.gameConfig.gui.bars.height, 
-                                                game.gameConfig.gui.bars.color.mp, true, 'mp', this);
-        
+        let bar2 = this.game.player.combat.createBar(this.game.cameras.main.width-this.game.gameConfig.gui.bars.width-1, 1, 
+                                                     this.game.gameConfig.gui.bars.width, this.game.gameConfig.gui.bars.height, 
+                                                     this.game.gameConfig.gui.bars.color.mp, true, 'mp', this);
+
+        this.game.rat.combat.createBar(0, 0, 8, 1, this.game.gameConfig.gui.bars.color.hp, false, 'hp', this);
         this.elements.push(bar, bar2);
         this.highlight = null;
-
+        this.highlighted = null;
+        
         // Create event listeners.
-        game.events.on('drawHighlight', function(obj, config) { return this.drawHighlight(obj, config); }, this)
-        game.events.on('destroyHighlight', function() { return this.destroyHighlight(); }, this)
+        this.game.events.on('drawHighlight', function(obj) { return this.highlighted = obj; }, this)
+        this.game.events.on('destroyHighlight', function() { return this.destroyHighlight(); }, this)
     }
 
     update()
@@ -36,32 +39,62 @@ class GUIScene extends Phaser.Scene {
         for(let i = 0; i < this.elements.length; i++){
             this.elements[i].draw();
         }
+
+        this.drawHighlight();
+
+        // Update coords/draw health bars of mobs.
+        for(let i = 0; i < this.game.actorsList.length; i++){
+            let combat = this.game.actorsList[i].combat;
+            if(combat){
+                if(combat.barHP){
+                    combat.barHP.x = Math.round(this.game.actorsList[i].x-4-this.game.cameras.main.worldView.x);
+                    combat.barHP.y = Math.round(this.game.actorsList[i].y-8-this.game.cameras.main.worldView.y);
+                    combat.barHP.draw();
+                }
+            } 
+        }
     }
 
-    drawHighlight(obj, config) {
-        // Grab scene to get camera coords.
-        let game = this.scene.get('GameScene');
-        
-        // Calculate actual x/y values offset by camera position.
-        let offsetX = Math.round((obj.x-(obj.sprite.width/2))-game.cameras.main.worldView.x);
-        let offsetY = Math.round((obj.y-(obj.sprite.height/2))-game.cameras.main.worldView.y);
+    drawHighlight() {
+        // Redraw every frame.
+        let config = this.game.gameConfig;
+        if(this.highlight){
+            this.highlight.text.destroy();
+            this.highlight.clear();
+        }
 
-        // Create a highlight rectangle around coords.
-        if(!this.highlight){
+        if(this.highlighted){
+            // Calculate actual x/y values offset by camera position.
+            let offsetX = Math.round((this.highlighted.x-(10))-this.game.cameras.main.worldView.x);
+            let offsetY = Math.round((this.highlighted.y-(10))-this.game.cameras.main.worldView.y);
+
+            // Create a highlight rectangle around coords.
             let graphics = this.add.graphics();
-            if(obj.walkable){
-                graphics.lineStyle(2, config.gui.highlights.walkable);
-            } else {
+            // Different colors depending on the status of the obj.
+            if(!this.highlighted.walkable){
                 graphics.lineStyle(2, config.gui.highlights.unwalkable);
+            } else if(this.highlighted.occupied){
+                graphics.lineStyle(2, config.gui.highlights.target);
+            } else {
+                graphics.lineStyle(2, config.gui.highlights.walkable);
             }
-            this.highlight = graphics.strokeRoundedRect(offsetX, offsetY, obj.sprite.width, obj.sprite.height, 4);
+            this.highlight = graphics.strokeRoundedRect(offsetX, offsetY, 20, 20, 7);
 
             // Draw objects name in bottom right corner of screen.
-            this.highlight.text = this.add.bitmapText(game.cameras.main.width-60, game.cameras.main.height, this.guiFont, obj.name).setAlpha(0.6);;
+            let text = this.highlighted.name
+            if(this.highlighted.occupied){
+                text = this.highlighted.occupied.name;
+                if(text == 'player'){
+                    text = this.highlighted.occupied.charName;
+                }
+            }
+
+            // Draw text.
+            this.highlight.text = this.add.bitmapText(this.game.cameras.main.width-60, this.game.cameras.main.height, this.guiFont, text).setAlpha(0.6);
             this.highlight.text.y -= (this.highlight.text.fontSize*1.5);
 
             // Set text color.
-            this.highlight.text.setTintFill(0xffffff);
+            this.highlight.text.setTintFill(config.gui.highlights.text);
         }
     }