+++ CHANGELOG.md
@@ -3,10 +3,26 @@
 ---
 # Etheria
 
+## 0.002 ()
+
+### Added
+- Added A* pathfinding.
+- New Mob: Mutated Rat
+- Added hp bars for mobs.
+- Added basic combat.
+- Added particles for when an actor takes damage.
+
+### Changed
+- Redid how the highlight rect is drawn to be less jittery.
+
+### Bug Fixes
+
+- Fixed the hp/mp bar text displaying at wrong coords.
+- Fixed the highlight text display at wrong coords.
 
 ## 0.001 (2019-4-20)
 
-### Content
+### Added
 
 - Initial changlog, not much here.
 - Player movement(will be reworked later)